requirements.txt
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==2.2.17
django-admin-sortable==2.2.3
django-allauth==0.41.0
django-debug-toolbar==3.2
django-widget-tweaks==1.4.8
idna==2.10
oauthlib==3.1.0
Pillow==2.2.2
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.2
